moved factes and hits dirs to searcher dir
