+++ instantsearch/src/commonMain/kotlin/com/algolia/instantsearch/helper/searcher/facets/FacetsSearcher.kt
@@ -1,10 +1,10 @@
-package com.algolia.instantsearch.helper.searcher.multi.facets
+package com.algolia.instantsearch.helper.searcher.facets
 
 import com.algolia.instantsearch.core.ExperimentalInstantSearch
 import com.algolia.instantsearch.core.searcher.Searcher
 import com.algolia.instantsearch.helper.searcher.SearcherScope
 import com.algolia.instantsearch.helper.searcher.multi.MultiSearcher
-import com.algolia.instantsearch.helper.searcher.multi.facets.internal.DefaultFacetsSearcher
+import com.algolia.instantsearch.helper.searcher.facets.internal.DefaultFacetsSearcher
 import com.algolia.instantsearch.helper.searcher.multi.internal.asMultiSearchComponent
 import com.algolia.search.client.ClientSearch
 import com.algolia.search.model.Attribute
